fig_dash/ui/apps/screenshot/requirements.txt
pyautogui
python-xlib